Fixed a type conversion issue in the windows affinity_t constructor error path

+++ src/affinity.hxx
@@ -102,7 +102,7 @@ namespace pcat
 					processorInfo = {count};
 				}
 				else
-					throw std::system_error{error, std::system_category()};
+					throw std::system_error{static_cast<int>(error), std::system_category()};
 			}
 
 			for (const auto &processor : processorInfo)
